math: port safe math functions from QQL core
Some of these are a bit frustrating, especially the `dist`
approximations, which are not only labeled wrong (upper and lower bounds
swapped) but also slower than just operating on the squared distance,
which cuts down the number of multiplications from 6 to 2 and the number
of square roots from 1 to 0 (in the worst case: it is also at least as
fast in every case).

But, since the entropy is frozen now, we're bound to them. Onward!