color: define color database types and parsing
It's time to get the canonical color data into the system. Defining all
the 153 colors and 7 palettes as code constants would be possible, but
unwieldy, especially given that the QQL JavaScript has a fair amount of
non-uniformity in how the colors are defined. Instead, we define a
simple database format that can be dumped with a script from the QQL
repo and read by this program at runtime. We also define an indexed
version of this database that keys color references by small integers
instead of strings.

On my laptop, it takes about 100μs to parse the JSON and 50μs to index
it into the database form. I experimented with using msgpack instead of
JSON, via `rmp-serde`: it gives both faster reads and smaller files,
but only by a factor of 2 or 3 (especially after gzip), so I suspect
that the broader compatibility of JSON is worth the cost here.

Of course, we could also go with a fancy Manish-style databake approach,
which wouldn't have the format compatibility issues but is probably just
more complexity than we need right now. The initialization cost should
only be paid once per binary invocation, and it's well under a
millisecond, so I'll call that good enough.

Color data dump built at commit 31e0414d7826 of the QQL art repo.