art: compute flow field vectors
This is spot-checked even less than usual, not just because I'm testing
only two points in the field but also because that probably doesn't test
whether disturbances are being computed correctly. For these structures
that are denser and harder to inspect, I'm planning to test things more
holistically once we get some graphics going.