art: add sectors helper structure
This isn't the next linear step in porting, but I kind of just wanted to
do it. I haven't copied over the ability for sectors to be lazily
populated; I think that that just won't be a notable performance
improvement, on either construction or iteration. This definitely
simplifies things; `affected` is much shorter and more straightforward
than its JavaScript counterpart.