bin: dump full point data from CLI
With this, we can directly compare point data between this port and the
canonical algorithm. Running on the seed for QQL #4, we can see that
about 95% of the points match exactly, while the remaining 5% have
errors in the last place or so of some floating point values. Here are
some deltas selected uniformly at random [1]:

```
{"bullseyeSpec":{"density":0.9266738796933677,"rings":3},"color":[43.8988694360886,82.2876822240302,96.26773205510689,100],"point":[225.09684579617493,674.7380070005676],"scale":1.8850861015035418,"secondaryColor":[40.298226097534155,4.489032426763748,94.3123832681131,100]}
{"bullseyeSpec":{"density":0.9266738796933677,"rings":3},"color":[43.8988694360886,82.2876822240302,96.26773205510689,100],"point":[225.09684579617496,674.7380070005676],"scale":1.8850861015035418,"secondaryColor":[40.298226097534155,4.489032426763748,94.3123832681131,100]}
// compare .point[0]: ...493 vs ...496

{"bullseyeSpec":{"density":0.9266738796933677,"rings":3},"color":[43.909770414642594,80,96.32958576490306,100],"point":[576.501726604539,442.7965164486189],"scale":1.8850861015035418,"secondaryColor":[149.52049244583475,58,35.59944288816344,100]}
{"bullseyeSpec":{"density":0.9266738796933677,"rings":3},"color":[43.909770414642594,80,96.32958576490306,100],"point":[576.5017266045392,442.7965164486189],"scale":1.8850861015035418,"secondaryColor":[149.52049244583475,58,35.59944288816344,100]}
// compare .point[0]: ...539 vs ...5392

{"bullseyeSpec":{"density":0.9266738796933677,"rings":3},"color":[42.1508350371102,85.53179581467471,98,100],"point":[518.5494780755229,1232.4889913670481],"scale":1.8850861015035418,"secondaryColor":[38,2.9422280478607585,95.65784324802262,100]}
{"bullseyeSpec":{"density":0.9266738796933677,"rings":3},"color":[42.1508350371102,85.53179581467471,98,100],"point":[518.5494780755229,1232.4889913670481],"scale":1.8850861015035418,"secondaryColor":[38,2.9422280478607576,95.65784324802262,100]}
// compare .secondaryColor[1]: ...585 vs ...576
```

These are actually different `f64` values, so I don't think that it's
likely to just be a difference in serialization, and I don't think that
values are being crushed through `f32` (they don't lie on the expected
boundaries). As the examples above show, differences can occur in both
position values and color values. The cause is not immediately obvious
to me, so I'm not sure whether it could in principle affect layout.

To investigate more later!

[1]: diff --suppress-common-lines -t -W600 -y /tmp/a /tmp/b | shuf -n 3 | tr '|' '\n' | sed -e 's/ *$//' -e 's/^ *//'