art: add fast-collisions and inflate-draw-radius options
These are extensions to the canonical algorithm. `--fast-collisions`
keeps the spirit of the algorithm but might slightly change the output
sometimes because approximations differ. `--inflate-draw-radius` causes
all points to have a minimum positive draw radius at paint time but not
at layout time (which matters for tiny or negative-scale points).