art: consider stroke thickness when culling viewport
We no longer incorrectly cull circles that would be out of the viewport
if the stroke had negligible thickness but are inside the viewport with
the actual stroke thickness.

For a test case, try this viewport into QQL #248:

    0x962a7d642178c87eeec92240669ec98f868dc54e4f6c97cf94ccffff12c9c923
    --viewport 0.005x0.00087+0.1225+0.204
    --width 100000

This output image is 500×109 px and contains just part of a circle.
Changing `0.00087` to `0.00081` shows an even smaller amount. In both
cases, before this patch the circle is culled incorrectly.