art: align chunks to pixel boundaries
It makes sense that this would be required. I made a silly error in my
earlier tests that suggested that this wasn't necessary (documented in
the now-deleted comment), wherein I wasn't testing the problematic case
at all because the chunk grid actually did fit evenly into the canvas
dimensions even though I picked funny numbers that are sure to not have
any large multiples, like "3".

This is a good change for lots of reasons. It improves performance
because compositing can always take the fast path of just blitting the
image onto the final canvas. It improves quality because we don't have
to do any anti-aliasing. And it improves flexibility because things
should be commutative now, so we can look into compositing more
aggressively (out of order) without losing determinism.

Also, wait, the order was non-deterministic anyway, based on the order
that we push to the vector. Sigh. :-)