sectors: fix panic on some very negative points
If `radius < 0.0` is small and `x_min == x_max` and `y_min == y_max`,
then the iteration will be over a single sector (the ranges are
inclusive). But if `radius < 0.0` is large such that `x_min == x_max`
but `y_min > y_max`, then the slicing will panic. In that case, we now
emulate the bare `for` loop from the original JavaScript implementation
by not iterating at all. We use a little extra bookkeeping to
deduplicate these calculations.

Seeds fixed by this:

0xe25ecd59478ce509232e37c2e0810fd4c655b8b4223cb2f742c6ffff10ea0917
0xe25ecd59478ce509232e37c2e0810fd4c655b8b4343952599a83ffff10ea0917
0xe25ecd59478ce509232e37c2e0810fd4c655b8b469e2b4dd0d07ffff10ea0917

(Thanks, `@heeey`!)

Seeds to cross-test here:

0xe25ecd59478ce509232e37c2e0810fd4c655b8b486bf1e348806ffff10e2cd33

(Should not be affected by this change, but can be broken with a naive
patch that causes *all* negative-radius points to be skipped entirely.)