art: retain insertion order of colors used
This way, it's equivalent to the JavaScript algorithm. This shouldn't do
much more work than previously, since the largest palette, Fidenza, has
only 16 colors. So the only appreciable impact should be using three
words of cache for the new `Vec<ColorKey>`.

Tested on QQL #29 and QQL #78, each of which uses every color, but in
different orders from each other. We're now consistent with the order
given by the JavaScript algorithm in both cases.