art: support rendering only to advance RNG
The `render` function (formerly called `draw`) now takes a type
parameter bounded by `PaintMode`. If this is `paint_mode::Paint`, then
the function renders as normal and returns a `DrawTarget`. If it is
`paint_mode::Skip`, then the function does not perform any graphics
operations and returns a `()`, but it still advances and mutates the
RNG. This lets callers quickly render a large set of points solely for
the purpose of learning what the RNG state will be after those points
are rendered.

Due to the way that the types are set up, it is impossible for `render`
to accidentally draw to the canvas in `paint_mode::Skip`. In fact, it
will never even allocate a canvas.

Test Plan:
Golden tests suffice.