art: make `(normal_points, splatter_sink)` more precise
The `AssertNone` variant is meant to be used only when there are no
normal points to render. We now require that at the type level to
prevent programming mistakes. As an added bonus, this makes the argument
value more self descriptive at the call site, as `NormalPoints::None`
instead of just a `&[]` that warranted a comment.